Feature/ac 4

+++ database/migrations/2022_03_04_204849_create_transactions_table.php
@@ -15,9 +15,9 @@ return new class extends Migration
     {
         Schema::create('transactions', function (Blueprint $table) {
             $table->id();
-            $table->integer('sender_wallet_id');
-            $table->integer('recipient_wallet_id');
-            $table->integer('volume');
+            $table->bigInteger('sender_wallet_id');
+            $table->bigInteger('recipient_wallet_id');
+            $table->bigInteger('volume');
             $table->integer('status');
             $table->integer('type');
             $table->timestamps();
